requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
